+++ Assets/Animations/New Folder.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 764edefdbc3f84e44abf6cce6862c8e2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
