+++ Assets/Animations/New Folder/Shot.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 04929f9a0c56ab4458b40514dda14aa3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
