+++ Assets/Animations/New Folder/ShotDown.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e9f1083b3cbc97341bfd11e11db5cd66
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
