+++ Assets/Animations/New Folder/ShotStraightUp.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4675a0a2307ec0d439ddb852a3372bf8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
