+++ Assets/Animations/New Folder/ShotUp.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4b1e63deb00981f4586146c3cd83c986
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
