+++ Assets/Animator/Test.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 11946486ae33f1449a2a98cf875bb5bd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
