--- Assets/Kevin Iglesias/Archer Animations/Scripts/BowLoadScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e2289541da6550c45bdcdbd207454893
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
