+++ Assets/Scripts/Minion/BowLoadScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2289541da6550c45bdcdbd207454893
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
