+++ Assets/Tex.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3ed36cee5b9c2d044bdce192caf09f87
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
